add hmp

+++ fastfold/utils/feats.py
@@ -44,7 +44,7 @@ def dgram_from_positions(
     )
     lower = torch.linspace(min_bin, max_bin, no_bins, device=pos.device) ** 2
     upper = torch.cat([lower[1:], lower.new_tensor([inf])], dim=-1)
-    dgram = ((dgram > lower) * (dgram < upper)).type(dgram.dtype)
+    dgram = ((dgram > lower).type(dgram.dtype) * (dgram < upper)).type(dgram.dtype)
 
     return dgram
 
@@ -91,7 +91,7 @@ def build_template_angle_feat(template_feats: Dict[str, Any]) -> torch.Tensor:
     torsion_angles_mask = template_feats["template_torsion_angles_mask"]
     template_angle_feat = torch.cat(
         [
-            nn.functional.one_hot(template_aatype, 22),
+            nn.functional.one_hot(template_aatype, 22).to(torch.float32),
             torsion_angles_sin_cos.reshape(*torsion_angles_sin_cos.shape[:-2], 14),
             alt_torsion_angles_sin_cos.reshape(
                 *alt_torsion_angles_sin_cos.shape[:-2], 14
@@ -136,10 +136,10 @@ def build_template_pair_feat(
     to_concat.append(
         aatype_one_hot[..., None, :, :].expand(
             *aatype_one_hot.shape[:-2], n_res, -1, -1
-        )
+        ).to(dgram.dtype)
     )
     to_concat.append(
-        aatype_one_hot[..., None, :].expand(*aatype_one_hot.shape[:-2], -1, n_res, -1)
+        aatype_one_hot[..., None, :].expand(*aatype_one_hot.shape[:-2], -1, n_res, -1).to(dgram.dtype)
     )
 
     n, ca, c = [rc.atom_order[a] for a in ["N", "CA", "C"]]
@@ -179,7 +179,7 @@ def build_template_pair_feat(
 def build_extra_msa_feat(batch: Dict[str, Any]) -> torch.Tensor:
     msa_1hot = nn.functional.one_hot(batch["extra_msa"], 23)
     msa_feat = [
-        msa_1hot,
+        msa_1hot.to(torch.float32),
         batch["extra_has_deletion"].unsqueeze(-1),
         batch["extra_deletion_value"].unsqueeze(-1),
     ]
